# CI speedup mamba/pip