--- a/.github/workflows/clean_notebooks.yml
+++ b/.github/workflows/clean_notebooks.yml
@@ -3,7 +3,7 @@ name: check notebooks
 on:
   pull_request:
     branches:
-    - master
+      - master
 
 jobs:
   check: