--- a/.github/workflows/pytest.yml
+++ b/.github/workflows/pytest.yml
@@ -25,12 +25,11 @@ jobs:
       - name: Setup Conda Environment
         uses: conda-incubator/setup-miniconda@v2
         with:
-          auto-update-conda: true
           environment-file: ./environment.yml
           activate-environment: weldx
           auto-activate-base: false
           use-only-tar-bz2: true # IMPORTANT: This needs to be set for caching to work properly!
-          channels: conda-forge, defaults
+          channels: conda-forge,defaults
           mamba-version: "*"
 
       - name: conda info