--- a/.github/workflows/pytest.yml
+++ b/.github/workflows/pytest.yml
@@ -48,3 +48,23 @@ jobs:
 
       - name: codecov.io
         uses: codecov/codecov-action@v1
+
+  pytest_pip:
+    name: pytest pip
+    runs-on: ${{ matrix.os }}
+    strategy:
+      matrix:
+        os: [ubuntu-18.04, ubuntu-latest, macos-latest, windows-latest]
+        py: [ '3.7', '3.8' ]
+    steps:
+      - uses: actions/checkout@v2
+      - uses: actions/setup-python@v2
+        with:
+          python-version: ${{ matrix.py }}
+      - name: pip installs
+        run: |
+          pip install .
+      - name: run pytest
+        run: |
+          pytest --runslow
+          coverage xml