--- a/.github/workflows/pytest.yml
+++ b/.github/workflows/pytest.yml
@@ -6,61 +6,39 @@ jobs:
     name: pytest
     runs-on: ${{ matrix.os }}
     strategy:
-      fail-fast: false
       matrix:
-        os: [ubuntu-18.04, ubuntu-latest, macos-latest, windows-latest]
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v2
-
-      - name: Cache conda
-        uses: actions/cache@v2
-        env:
-          # Increase this value to reset cache if ./environment.yml has not changed
-          CACHE_NUMBER: 0
-        with:
-          path: ~/conda_pkgs_dir
-          key: ${{ runner.os }}-conda-${{ env.CACHE_NUMBER }}-${{ hashFiles('./environment.yml') }}
-
-      - name: Setup Conda Environment
-        uses: conda-incubator/setup-miniconda@v2
-        with:
-          environment-file: ./environment.yml
-          activate-environment: weldx
-          auto-activate-base: false
-          use-only-tar-bz2: true # IMPORTANT: This needs to be set for caching to work properly!
-          channels: conda-forge,defaults
-          mamba-version: "*"
-
-      - name: conda info
-        shell: bash -l {0}
-        run: conda info
-
-      - name: conda list
-        shell: bash -l {0}
-        run: conda list
-
-      - name: pytest
-        shell: bash -l {0}
-        run: |
-          pytest --runslow
-          coverage xml
-
-      - name: codecov.io
-        uses: codecov/codecov-action@v1
-
-  pytest_pip:
-    name: pytest pip
-    runs-on: ${{ matrix.os }}
-    strategy:
-      matrix:
-        os: [ubuntu-18.04, ubuntu-latest, macos-latest, windows-latest]
-        py: [ '3.7', '3.8' ]
+        os: [ubuntu-latest, macos-latest, windows-latest]
+        py: [ '3.7' ]
     steps:
       - uses: actions/checkout@v2
       - uses: actions/setup-python@v2
         with:
           python-version: ${{ matrix.py }}
+
+      - uses: actions/cache@v2
+        if: startsWith(runner.os, 'ubuntu')
+        with:
+          path: ~/.cache/pip
+          key: ${{ runner.os }}-pip-${{ hashFiles('setup.cfg') }}
+          restore-keys: |
+            ${{ runner.os }}-pip-
+
+      - uses: actions/cache@v2
+        if: startsWith(runner.os, 'macOS')
+        with:
+          path: ~/Library/Caches/pip
+          key: ${{ runner.os }}-pip-${{ hashFiles('setup.cfg') }}
+          restore-keys: |
+            ${{ runner.os }}-pip-
+
+      - uses: actions/cache@v2
+        if: startsWith(runner.os, 'Windows')
+        with:
+          path: ~\AppData\Local\pip\Cache
+          key: ${{ runner.os }}-pip-${{ hashFiles('setup.cfg') }}
+          restore-keys: |
+            ${{ runner.os }}-pip-
+
       - name: pip installs
         run: |
           pip install pytest pytest-cov setuptools_scm
@@ -69,3 +47,6 @@ jobs:
         run: |
           pytest --runslow
           coverage xml
+
+      - name: codecov.io
+        uses: codecov/codecov-action@v1