--- a/.github/workflows/pytest.yml
+++ b/.github/workflows/pytest.yml
@@ -8,7 +8,7 @@ jobs:
     strategy:
       matrix:
         os: [ubuntu-latest, macos-latest, windows-latest]
-        py: [ '3.7' ]
+        py: ['3.7']
     steps:
       - uses: actions/checkout@v2
       - uses: actions/setup-python@v2