--- a/.github/workflows/pytest.yml
+++ b/.github/workflows/pytest.yml
@@ -8,7 +8,7 @@ jobs:
     strategy:
       matrix:
         os: [ubuntu-latest, macos-latest, windows-latest]
-        py: ['3.7', '3.8', '3.9']
+        py: ['3.7', '3.8']
     steps:
       - uses: actions/checkout@v2
       - uses: actions/setup-python@v2