--- a/.github/workflows/pytest_asdf.yml
+++ b/.github/workflows/pytest_asdf.yml
@@ -8,7 +8,7 @@ jobs:
     strategy:
       matrix:
         os: [ubuntu-18.04]
-        py: [ '3.7', '3.8' ]
+        py: ['3.7', '3.8']
     steps:
       - uses: actions/checkout@v2
       - uses: actions/setup-python@v1