--- a/.github/workflows/trigger_binder_build.yml
+++ b/.github/workflows/trigger_binder_build.yml
@@ -6,7 +6,7 @@ on:
 
 jobs:
   trigger-binder-build:
-    runs-on: [ubuntu-latest]
+    runs-on: [ ubuntu-latest ]
     steps:
       - uses: s-weigand/trigger-mybinder-build@v1
         with: