--- a/doc/rtd_environment.yml
+++ b/doc/rtd_environment.yml
@@ -1,0 +1,19 @@
+channels:
+  - conda-forge
+  - defaults
+dependencies:
+  - python=3.7
+  - setuptools_scm
+  # documentation
+  - sphinx
+  - recommonmark
+  - sphinxcontrib-napoleon
+  - nbsphinx
+  - sphinx-copybutton
+  - pydata-sphinx-theme
+  - numpydoc>=0.5
+  - sphinx-autodoc-typehints
+  # pip packages
+  - pip
+  - pip:
+    - git+https://github.com/CagtayFabry/sphinx-asdf.git@sphinx-weldx