--- a/doc/rtd_environment.yml
+++ b/doc/rtd_environment.yml
@@ -1,3 +1,4 @@
+name: weldx
 channels:
   - conda-forge
   - defaults