--- a/environment.yml
+++ b/environment.yml
@@ -40,7 +40,6 @@ dependencies:
   - sphinxcontrib-napoleon
   - nbsphinx
   - sphinx-copybutton
-  - sphinxcontrib-bibtex
   - pydata-sphinx-theme
   - numpydoc>=0.5
   - sphinx-autodoc-typehints