--- a/setup.cfg
+++ b/setup.cfg
@@ -51,6 +51,7 @@ install_requires =
     Jinja2 >=2.11
     networkx >=2
     matplotlib >=3
+    ipywidgets
 include_package_data = True
 
 [options.entry_points]