## Changes
I ended up adding a few improvements to speed up the pipelines:
- pytest setup installs via pip (installs ~1-2 min on all OS)
- pytests runs on python matrix (currently 3.7 & 3.8)
- skip the auto conda update step (one less time solving conda envs)
- use mamba for conda environment setup in remaining conda actions (a little bit faster)
- ReadTheDocs now has it's own (stripped down) environment file `doc/rtd_environment.yml`. Conda build time improved from >400s to 150s 🥳 . Pip installs only increased from 40s to 70s so quite a net benefit here
- I added `ipywidgets` to the requirement in `setup.cfg` since some tutorials need them (CSM). We should keep this in mind once we split the packages further

## Related Issues
Closes # (add issue numbers)

## Checks
- [x] updated CHANGELOG.md
- [x] updated tests
- [x] updated doc/
- [x] update example/tutorial notebooks