@CagtayFabry 2020-12-15
no idea what happened [here](https://readthedocs.org/projects/weldx/builds/12568494/) but it's a sight to behold @vhirtham 😆