@CagtayFabry 2020-12-16
see PR notes for detail @vhirtham 

We will need to update the GitHub merge protection rules since the pytest action names changed slightly (including python version)